% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_all}
\alias{get_all}
\title{Lookup for IP address geolocation and proxy information}
\usage{
get_all(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return all the geolocation and proxy information about the IP address
}
\description{
Lookup for a comprehensive of the information such as location, ASN, ISP and proxy. The availability of the data will depends on the plan you have signed up for your API key.
}
\examples{
\dontrun{
get_all("1.0.241.135")
}

}
