% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Locationio.R
\name{lookupCoordinateByIP}
\alias{lookupCoordinateByIP}
\title{Lookup an IP address's coordinate}
\usage{
lookupCoordinateByIP(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return the coordinate of the the IP address
}
\description{
Lookup for the IP address's coordinate
}
\examples{
\dontrun{
lookupCoordinateByIP("1.0.241.135")
}

}
