% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Locationio.R
\name{lookupCountryByIP}
\alias{lookupCountryByIP}
\title{Lookup an IP address's country}
\usage{
lookupCountryByIP(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return the country name of the the IP address
}
\description{
Lookup for the IP address's country
}
\examples{
\dontrun{
lookupCountryByIP("1.0.241.135")
}

}
