% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Locationio.R
\name{lookupHostedDomain}
\alias{lookupHostedDomain}
\title{Lookup for domains hosted on an IP address.}
\usage{
lookupHostedDomain(ip, page)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}

\item{page}{(optional) Pagination result returns of the hosted domains. If unspecified, 1st page will be used.}
}
\value{
Return list of hosted domain names by the IP address
}
\description{
Lookup for a list of hosted domain names by the IP address.
}
\examples{
\dontrun{
lookupHostedDomain("1.0.241.135")
}

}
