% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Locationio.R
\name{lookupLocationByIP}
\alias{lookupLocationByIP}
\title{Lookup an IP address's location in text}
\usage{
lookupLocationByIP(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return the location of the the IP address in text
}
\description{
Lookup for the IP address's location in text
}
\examples{
\dontrun{
lookupLocationByIP("1.0.241.135")
}

}
