% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Locationio.R
\name{setApiKey}
\alias{setApiKey}
\title{Set IP2Location.io API key}
\usage{
setApiKey(api_key)
}
\arguments{
\item{api_key}{IP2Location.io API key}
}
\description{
Set IP2Location.io API key for lookup. Free API key can be obtained from <https://www.ip2location.io/sign-up?ref=1/>
}
\examples{
\dontrun{
setApiKey("YOUR_API_KEY")
}

}
